Back

+++ src/schema/product/product.resolvers.js
@@ -1,3 +1,4 @@
+const { GraphQLError } = require('graphql');
 const ProductService = require('../../services/product.service');
 
 const productResolvers = {
@@ -17,14 +18,36 @@ const productResolvers = {
 
     // MUTATIONS
     Mutation: {
-        addProduct: async (parent, args) => {
+        addProduct: async (parent, args, context) => {
+
+            if (!context.user) {
+                throw new GraphQLError('Debes estar autenticado para crear productos.', {
+                    extensions: { code: 'UNAUTHENTICATED' }
+                });
+            }
             return await ProductService.createProduct(args);
         },
-        updateProduct: async (parent, { id, ...updates }) => {
+        updateProduct: async (parent, { id, ...updates }, context) => {
+            if (!context.user) {
+                throw new GraphQLError('Debes estar autenticado para actualizar productos.', {
+                    extensions: { code: 'UNAUTHENTICATED' }
+                });
+            }
             // Desestructuramos para separar el ID de los datos a actualizar
             return await ProductService.updateProduct(id, updates);
         },
-        deleteProduct: async (parent, args) => {
+        deleteProduct: async (parent, args, context) => {
+            if (!context.user) {
+                throw new GraphQLError('No estás autenticado.', {
+                    extensions: { code: 'UNAUTHENTICATED' }
+                });
+            }
+
+            if (context.user.role !== 'ADMIN') {
+                throw new GraphQLError('No tienes permisos para eliminar productos.', {
+                    extensions: { code: 'FORBIDDEN' }
+                })
+            }
             return await ProductService.deleteProduct(args.id);
         }
     }
